Move emulated session from core/ to shell/
