+++ cowrie/commands/adduser.py
@@ -7,7 +7,7 @@ import random
 
 from twisted.internet import reactor
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
