+++ cowrie/commands/dd.py
@@ -9,8 +9,8 @@ from __future__ import division, absolute_import
 
 from os import path
 
-from cowrie.core.honeypot import HoneyPotCommand
-from cowrie.core.fs import *
+from cowrie.shell.honeypot import HoneyPotCommand
+from cowrie.shell.fs import *
 
 commands = {}
