+++ cowrie/commands/env.py
@@ -3,7 +3,7 @@ from __future__ import division, absolute_import
 
 from twisted.internet import reactor
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
