+++ cowrie/commands/free.py
@@ -9,7 +9,7 @@ from __future__ import division, absolute_import
 
 import getopt
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
