+++ cowrie/commands/fs.py
@@ -11,8 +11,8 @@ from __future__ import division, absolute_import
 import getopt
 import copy
 from os import path
-from cowrie.core.honeypot import HoneyPotCommand
-from cowrie.core.fs import *
+from cowrie.shell.honeypot import HoneyPotCommand
+from cowrie.shell.fs import *
 
 commands = {}
