+++ cowrie/commands/gcc.py
@@ -10,7 +10,7 @@ import os
 
 from twisted.internet import reactor
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
