+++ cowrie/commands/last.py
@@ -3,8 +3,8 @@
 
 from __future__ import division, absolute_import
 
-from cowrie.core.honeypot import HoneyPotCommand
-from cowrie.core.fs import *
+from cowrie.shell.honeypot import HoneyPotCommand
+from cowrie.shell.fs import *
 from cowrie.core import utils
 
 commands = {}
