+++ cowrie/commands/nohup.py
@@ -7,8 +7,8 @@ from __future__ import division, absolute_import
 import os
 import getopt
 
-from cowrie.core.honeypot import HoneyPotCommand
-from cowrie.core.fs import *
+from cowrie.shell.honeypot import HoneyPotCommand
+from cowrie.shell.fs import *
 
 commands = {}
