+++ cowrie/commands/perl.py
@@ -11,7 +11,7 @@ import getopt
 
 from twisted.python import log
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
