+++ cowrie/commands/ping.py
@@ -11,7 +11,7 @@ import getopt
 
 from twisted.internet import reactor
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
