+++ cowrie/commands/sleep.py
@@ -9,7 +9,7 @@ from __future__ import division, absolute_import
 
 from twisted.internet import reactor
 
-from cowrie.core.honeypot import HoneyPotCommand
+from cowrie.shell.honeypot import HoneyPotCommand
 
 commands = {}
