+++ cowrie/shell/customparser.py
