+++ cowrie/shell/fs.py
