+++ cowrie/shell/pwd.py
