+++ cowrie/shell/shlex.py
